chore: update hero images
