+++ public/assets/images/index.ts
@@ -16,11 +16,11 @@ import footerLogo from "./footer-logo.png";
 
 import offer from "./offer.svg";
 
-import shoe4 from "./shoe4.svg";
-import shoe5 from "./shoe5.svg";
-import shoe6 from "./shoe6.svg";
-import shoe7 from "./shoe7.svg";
-import shoe8 from "./shoe8.svg";
+import twinstroller from "./twin-stroller.png";
+import twoInOneStroller from "./two-in-one-stroller.png";
+
+import walker1 from "./walker1.png";
+import walker2 from "./walker2.png";
 
 import thumbnailBackground from "./thumbnail-background.svg";
 import thumbnailShoe1 from "./thumbnail-shoe1.svg";
@@ -40,11 +40,10 @@ export {
   companyLogo,
   footerLogo,
   offer,
-  shoe4,
-  shoe5,
-  shoe6,
-  shoe7,
-  shoe8,
+  twinstroller,
+  twoInOneStroller,
+  walker1,
+  walker2,
   thumbnailBackground,
   thumbnailShoe1,
   thumbnailShoe2,
